Move Error and ParseFailure into a separate module
