+++ src/error.rs
@@ -1,0 +1,101 @@
+use crate::item::Item;
+
+/// Unsuccessful command line parsing outcome, internal representation
+#[derive(Debug)]
+pub enum Error {
+    /// Parsing failed, it is still possible to improve the error message
+    /// bool shows if this error can be caught with fallback, messages with it set to true are
+    /// generated by .some(msg) modifier and allow to use it with .fallback
+    Message(String, bool),
+    /// Parsing failed and this is the final output
+    ParseFailure(ParseFailure),
+    /// Expected one of those values
+    ///
+    /// Used internally to generate better error messages
+    Missing(Vec<Item>),
+}
+
+impl Error {
+    #[must_use]
+    pub(crate) fn combine_with(self, other: Self) -> Self {
+        #[allow(clippy::match_same_arms)]
+        match (self, other) {
+            // help output takes priority
+            (a @ Error::ParseFailure(_), _) => a,
+            (_, b @ Error::ParseFailure(_)) => b,
+
+            // parsing failure takes priority
+            (a @ Error::Message(_, _), _) => a,
+            (_, b @ Error::Message(_, _)) => b,
+
+            // combine missing elements
+            (Error::Missing(mut a), Error::Missing(mut b)) => {
+                a.append(&mut b);
+                Error::Missing(a)
+            }
+        }
+    }
+}
+
+/// Unsuccessful command line parsing outcome, use it for unit tests
+///
+/// Useful for unit testing for user parsers, consume it with
+/// [`ParseFailure::unwrap_stdout`] and [`ParseFailure::unwrap_stdout`]
+#[derive(Clone, Debug)]
+pub enum ParseFailure {
+    /// Print this to stdout and exit with success code
+    Stdout(String),
+    /// Print this to stderr and exit with failure code
+    Stderr(String),
+}
+
+impl ParseFailure {
+    /// Returns the contained `stderr` values - for unit tests
+    ///
+    /// # Panics
+    ///
+    /// Panics if failure contains `stdout`
+    #[allow(clippy::must_use_candidate)]
+    #[track_caller]
+    pub fn unwrap_stderr(self) -> String {
+        match self {
+            Self::Stderr(err) => err,
+            Self::Stdout(_) => {
+                panic!("not an stderr: {:?}", self)
+            }
+        }
+    }
+
+    /// Returns the contained `stdout` values - for unit tests
+    ///
+    /// # Panics
+    ///
+    /// Panics if failure contains `stderr`
+    #[allow(clippy::must_use_candidate)]
+    #[track_caller]
+    pub fn unwrap_stdout(self) -> String {
+        match self {
+            Self::Stdout(err) => err,
+            Self::Stderr(_) => {
+                panic!("not an stdout: {:?}", self)
+            }
+        }
+    }
+
+    /// Run an action appropriate to the failure and produce the exit code
+    ///
+    /// Prints a message to `stdout` or `stderr` and returns the exit code
+    #[allow(clippy::must_use_candidate)]
+    pub fn exit_code(self) -> i32 {
+        match self {
+            ParseFailure::Stdout(msg) => {
+                print!("{}", msg); // completions are sad otherwise
+                0
+            }
+            ParseFailure::Stderr(msg) => {
+                eprintln!("{}", msg);
+                1
+            }
+        }
+    }
+}
