+++ src/info.rs
@@ -2,47 +2,9 @@
 
 use crate::{
     args::{Args, Conflict},
-    item::Item,
     parsers::ParseCommand,
-    ParseFailure, Parser,
+    Error, ParseFailure, Parser,
 };
-
-/// Unsuccessful command line parsing outcome, internal representation
-#[derive(Debug)]
-pub enum Error {
-    /// Parsing failed, it is still possible to improve the error message
-    /// bool shows if this error can be caught with fallback, messages with it set to true are
-    /// generated by .some(msg) modifier and allow to use it with .fallback
-    Message(String, bool),
-    /// Parsing failed and this is the final output
-    ParseFailure(ParseFailure),
-    /// Expected one of those values
-    ///
-    /// Used internally to generate better error messages
-    Missing(Vec<Item>),
-}
-
-impl Error {
-    #[must_use]
-    pub(crate) fn combine_with(self, other: Self) -> Self {
-        #[allow(clippy::match_same_arms)]
-        match (self, other) {
-            // help output takes priority
-            (a @ Error::ParseFailure(_), _) => a,
-            (_, b @ Error::ParseFailure(_)) => b,
-
-            // parsing failure takes priority
-            (a @ Error::Message(_, _), _) => a,
-            (_, b @ Error::Message(_, _)) => b,
-
-            // combine missing elements
-            (Error::Missing(mut a), Error::Missing(mut b)) => {
-                a.append(&mut b);
-                Error::Missing(a)
-            }
-        }
-    }
-}
 
 /// Information about the parser
 ///
